tests/requirements.txt
pytest
Faker
requests
allure-pytest
requests_to_curl
